+++ test/unit/fuzz-tools.test.js
@@ -243,6 +243,7 @@ import * as obgynV250 from '../../lib/obgyn-v250.js';
 import * as cardiometabV251 from '../../lib/cardiometab-v251.js';
 import * as orthospineV252 from '../../lib/orthospine-v252.js';
 import * as radmeasureV253 from '../../lib/radmeasure-v253.js';
+import * as enturopsychV254 from '../../lib/enturopsych-v254.js';
 
 const MODULES = {
   'clinical.js': clinical,
@@ -430,6 +431,7 @@ const MODULES = {
   'cardiometab-v251.js': cardiometabV251,
   'orthospine-v252.js': orthospineV252,
   'radmeasure-v253.js': radmeasureV253,
+  'enturopsych-v254.js': enturopsychV254,
 };
 
 const MATRIX = [0, -1, 1e9, NaN, Infinity, -Infinity, '', undefined, null];